c8.5

+++ src/main/java/com/example/gestionEntreprise/model/CandidateSkill.java
@@ -8,7 +8,7 @@ public class CandidateSkill {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "idcandidate_skill")
-    private Long idCandidateSkill;
+    private Integer idCandidateSkill; // Changé de Long à Integer
 
     @ManyToOne
     @JoinColumn(name = "idcandidate")
@@ -23,8 +23,8 @@ public class CandidateSkill {
     private Level level;
 
     // Getters and setters
-    public Long getIdCandidateSkill() { return idCandidateSkill; }
-    public void setIdCandidateSkill(Long idCandidateSkill) { this.idCandidateSkill = idCandidateSkill; }
+    public Integer getIdCandidateSkill() { return idCandidateSkill; }
+    public void setIdCandidateSkill(Integer idCandidateSkill) { this.idCandidateSkill = idCandidateSkill; }
     public Candidate getCandidate() { return candidate; }
     public void setCandidate(Candidate candidate) { this.candidate = candidate; }
     public Skill getSkill() { return skill; }
