+++ src/main/java/com/example/gestionEntreprise/model/Degree.java
@@ -8,14 +8,14 @@ public class Degree {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "iddegree")
-    private Long idDegree; // Changé de Integer à Long
+    private Integer idDegree; // Changé de Long à Integer
 
     @Column(name = "degree_name", nullable = false, length = 100)
     private String degreeName;
 
     // Getters and setters
-    public Long getIdDegree() { return idDegree; }
-    public void setIdDegree(Long idDegree) { this.idDegree = idDegree; }
+    public Integer getIdDegree() { return idDegree; }
+    public void setIdDegree(Integer idDegree) { this.idDegree = idDegree; }
     public String getDegreeName() { return degreeName; }
     public void setDegreeName(String degreeName) { this.degreeName = degreeName; }
 }
