+++ src/main/java/com/example/gestionEntreprise/model/EducationLevel.java
@@ -8,14 +8,14 @@ public class EducationLevel {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "ideducation_level")
-    private Long idEducationLevel; // Changé de Integer à Long
+    private Integer idEducationLevel; // Changé de Long à Integer
 
     @Column(name = "level_name", nullable = false, length = 100)
     private String levelName;
 
     // Getters and setters
-    public Long getIdEducationLevel() { return idEducationLevel; }
-    public void setIdEducationLevel(Long idEducationLevel) { this.idEducationLevel = idEducationLevel; }
+    public Integer getIdEducationLevel() { return idEducationLevel; }
+    public void setIdEducationLevel(Integer idEducationLevel) { this.idEducationLevel = idEducationLevel; }
     public String getLevelName() { return levelName; }
     public void setLevelName(String levelName) { this.levelName = levelName; }
 }
