+++ src/main/java/com/example/gestionEntreprise/repository/CandidateRepository.java
@@ -5,12 +5,15 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface CandidateRepository extends JpaRepository<Candidate, Long> {
+public interface CandidateRepository extends JpaRepository<Candidate, Integer> {
     @Query("SELECT c FROM Candidate c JOIN c.person p WHERE p.email = :email")
     Candidate findByPersonEmail(@Param("email") String email);
 
     @Query("SELECT c FROM Candidate c WHERE c.person.idPerson = :personId")
-    Candidate findByPersonId(@Param("personId") Long personId);
+    Candidate findByPersonId(@Param("personId") Integer personId);
+
+
+    
 }
