+++ src/main/java/com/example/gestionEntreprise/repository/DegreeRepository.java
@@ -4,5 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.gestionEntreprise.model.Degree;
 
-public interface DegreeRepository extends JpaRepository<Degree, Long> {
+public interface DegreeRepository extends JpaRepository<Degree, Integer> 
+{
 }
