+++ src/main/java/com/example/gestionEntreprise/repository/EducationLevelRepository.java
@@ -4,5 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.gestionEntreprise.model.EducationLevel;
 
-public interface EducationLevelRepository extends JpaRepository<EducationLevel, Long> {
+public interface EducationLevelRepository extends JpaRepository<EducationLevel, Integer>
+ {
+
+    
 }
