+++ src/main/java/com/example/gestionEntreprise/repository/SkillRepository.java
@@ -4,5 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.gestionEntreprise.model.Skill;
 
-public interface SkillRepository extends JpaRepository<Skill, Long> {
+public interface SkillRepository extends JpaRepository<Skill, Integer>
+ {
 }
